extend event emitters and socket.io with history
Added new socket.withHistory.on and eventEmitter.withHistory.on
methods.  Causes the object to keep a 1 level cache of all events.
When a function registers for the event with the on method, the last
message that was cached is sent.  This is intended to remove the boiler
plate code where new objects have to both wire up for message updates
and make a initial request to get the latest state.  Now by wiring up
for message updates they automatically get sent the latest
state/message.