Peer Viewing enabled
If another person tries to connect to an ROV, they will get redirected
to connect peer-2-peer with the existing pilot in a readonly mode.
When the existing pilot quits, the browser will refresh and a new
person will take over as pilot.  A browser can be forced to always
connect directly by using force=true in the URL parameters