Background sync (#108)
Initial release of background-sync

This feature adds a "SYNC" button to the dives applet.  This sync feature will upload all dive data to openrov servers for backup and web based replay of dives.  The system has a status bar for tracking the uploads.  These uploads will take place even if the browser is closed.  

The user must be logged in with an OpenROV cloud account for the sync to work.  The account must be provisioned to work.