MJPG socket.io and http delivery work side by side (#132)
Refactored the socket.io based mjpeg player from the video.js to a stand alone web component.
Removed some of the code complexity around worker processes as it was not clear that it was actually fully implemented and giving the desired outcome.
Fixed some key code that was commented out in the camera registration message flow
Made sure that http based mjpeg cameras only load when http sources exist
Performance:
By disabling and re-enabling the 30fps 640x480 stream I saw a 10% cpu hit in the 'program' category. None of its routines are in the top 10.